:bug: Remove usage of bundled libc++ for mac
Attempting to use the bundled libc++ and libc++abi from mac with this
conan recipe has only resulted in strange runtime failures. For example
strong_ptr's demos crashed on exit() due to what I believe to be an
memory corruption bug due to an ABI conflict between the LLVM's bundled
dynamic libraries and the system ones. I'd prefer to use the libraries
bundled with LLVM to get the most out of the compiler, but alas, it
seems thats not possible. And I'm not sure if more time spent trying
to make this work will be fruitful.